refactor(ui): extract provider filter mapping to reusable utility
